% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{.options}
\alias{.options}
\title{VSD Options (Internal)}
\usage{
.options(arguments = list(), ...)
}
\arguments{
\item{arguments}{Graph-specific arguments}

\item{...}{General arguments}
}
\value{
List of list of graph options
}
\description{
Generates lists of lists of options for each graph type, from function call
}
\keyword{internal}
