% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{.subset_options}
\alias{.subset_options}
\title{VSD (Sub)Options (Internal)}
\usage{
.subset_options(subset, ellipsis, subarguments = NULL)
}
\arguments{
\item{subset}{Graph-specific preset (and allowed) values}

\item{ellipsis}{General arguments}

\item{subarguments}{Graph-specific arguments}
}
\value{
subset
}
\description{
Agglutinates preset, ellipsis, and arguments under a graph type
}
\keyword{internal}
