% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{get_options}
\alias{get_options}
\title{Title}
\usage{
get_options(object, prepend = list(), arguments, ...)
}
\arguments{
\item{object}{String identifying the type of graph}

\item{prepend}{Additional arguments to put at the start of the list}

\item{arguments}{Graph-specific arguments (from function call)}

\item{...}{General arguments (from function call)}
}
\value{

}
\description{
Title
}
