% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.r
\name{vsd}
\alias{vsd}
\title{Visualizing Survival Data}
\usage{
vsd(data = c())
}
\arguments{
\item{data}{The input data}
}
\description{
This function tries to guess the best format to render the inputted survival data analysis data into a graphically pleasing output, while also explaining any data processing that has occured during the way.
Or it will one day, for now it only goes "Bloop!"
}
\examples{
whoded()
}
