% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.r
\name{vsd}
\alias{vsd}
\title{Visualizing Survival Data}
\usage{
vsd(fit, data = NULL, ...)
}
\arguments{
\item{fit}{The estimated survival curve for the model, defaults to a single}

\item{data}{Dataframe from where the model fetches its variable, if left blank will be extracted from the model, if possible}

\item{...}{Kaplain-Meyer estimation of \code{"surv"}}
}
\value{
A list of graphs, relevant to the model
}
\description{
This function tries to guess the best format to render the inputted survival
data analysis data into a graphically pleasing output, while also explaining
any data processing that has occured during the way. Or it will one day, for
now it only goes \code{"plot()"}.
}
\examples{
# non-models are cohersed into a survfit with default arguments
\dontrun{
vsd(with(aml, Surv(time, status)))
vsd(coxph(Surv(time, status) ~ sex + rx + adhere, data = colon))
}

# survival fit model
vsd(survfit(Surv(futime, fustat) ~ rx, data = ovarian))

# coxph (with and without strata)
vsd(survfit(coxph(Surv(time, status) ~ sex + rx + adhere, data = colon)), colon)
vsd(survfit(coxph(Surv(time, status) ~ sex + strata(rx) + adhere, data = colon)), colon)

# parametric models
vsd(flexsurv::flexsurvreg(Surv(rectime, censrec) ~ group, data = flexsurv::bc, dist = "gengamma"))
}
