% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vsd.R
\name{vsd}
\alias{vsd}
\alias{vsd.formula}
\alias{vsd.Surv}
\alias{vsd.coxph}
\alias{vsd.survfitcox}
\alias{vsd.survfit}
\alias{vsd.flexsurvreg}
\title{Visualizing Survival Data}
\usage{
vsd(
  fit,
  data = NULL,
  arguments = list(),
  interactive = FALSE,
  .include = c("fit", "parametric", "forest", "residuals", "hazard"),
  ...
)

\method{vsd}{formula}(
  fit,
  data = NULL,
  arguments = list(),
  interactive = FALSE,
  .include = c("fit", "hazard"),
  ...
)

\method{vsd}{Surv}(
  fit,
  data = NULL,
  arguments = list(),
  interactive = FALSE,
  .include = c("fit", "hazard"),
  ...
)

\method{vsd}{coxph}(
  fit,
  data = NULL,
  arguments = list(),
  interactive = FALSE,
  .include = c("fit", "forest", "residuals", "hazard"),
  ...
)

\method{vsd}{survfitcox}(
  fit,
  data = NULL,
  arguments = list(),
  interactive = FALSE,
  .include = c("fit", "forest", "residuals", "hazard"),
  ...
)

\method{vsd}{survfit}(
  fit,
  data = NULL,
  arguments = list(),
  interactive = FALSE,
  .include = c("fit", "hazard"),
  ...
)

\method{vsd}{flexsurvreg}(
  fit,
  data = NULL,
  arguments = list(),
  interactive = FALSE,
  .include = c("fit", "parametric", "hazard"),
  ...
)
}
\arguments{
\item{fit}{The model, or Surv object, to fit the survival data}

\item{data}{Dataframe from where the model fetches its variables, if left
blank will be extracted from the model, if possible}

\item{arguments}{Collection of list of arguments, indexed by the specific
type of graph they should be passed to, has priority over \dots}

\item{interactive}{Allows to explore the generated graphs before returning
(use with \code{'plotly'} for best results)}

\item{.include}{Graph types to output if relevant, defaults to all possible}

\item{...}{Miscellaneous arguments, passed to ALL graphs}
}
\value{
A list of graphs, relevant to the model
}
\description{
This function tries to guess the best format to render the inputted survival
data analysis data into a graphically pleasing output.
}
\details{
More info will come here soon.
}
\section{Methods (by class)}{
\itemize{
\item \code{formula}: Wraps \code{Surv ~ (...)} in a survfit object (Kaplan-Meier
model)

\item \code{Surv}: \code{Surv(...)} in a survfit object (Kaplan-Meier
model)

\item \code{coxph}: Wraps \code{coxph(...)} in a survfit object (Kaplan-Meier
model)

\item \code{survfitcox}: Graphical output for survfit objects (Cox model)

\item \code{survfit}: Graphical output for survfit objects (Kaplan-Meier model)

\item \code{flexsurvreg}: Graphical output for flexsurvreg objects (parametric models)
}}

\examples{
# non-models are cohersed into a survfit object with default arguments
vsd(with(aml, Surv(time, status)))
vsd(Surv(time, status) ~ ph.ecog, data=lung)

# survival fit model
vsd(survfit(Surv(futime, fustat) ~ rx, data = ovarian))

# coxph (with and without strata)
vsd(coxph(Surv(time, status) ~ sex + rx + adhere, data = colon))
vsd(survfit(coxph(Surv(time, status) ~ sex + strata(rx) + adhere, data = colon)))

# parametric models
vsd(flexsurv::flexsurvreg(Surv(rectime, censrec) ~ group, data = flexsurv::bc, dist = 'gengamma'))
}
