% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.r
\name{vsd}
\alias{vsd}
\title{Visualizing Survival Data}
\usage{
vsd(fit, data = NULL, ...)
}
\arguments{
\item{fit}{The estimated survival curve for the model, defaults to a single
Kaplain-Meyer estimation of \code{"surv"}}

\item{\dots}{Optional arguments to be passed to graphic generation}

\item{surv}{The input Survival object, from \code{"Surv()"}}
}
\description{
This function tries to guess the best format to render the inputted survival
data analysis data into a graphically pleasing output, while also explaining
any data processing that has occured during the way. Or it will one day, for
now it only goes \code{"plot()"}.
}
\examples{
# non-models are cohersed into a survfit with default arguments
vsd(with(aml, Surv(time, status)))
vsd(coxph(Surv(time, status) ~ sex + rx + adhere, data = colon))

# example of models (with a repeated coxph)
vsd(survfit(Surv(futime, fustat) ~ rx, data = ovarian))
vsd(survfit(coxph(Surv(time, status) ~ sex + rx + adhere, data = colon)), colon)
vsd(flexsurv::flexsurvreg(Surv(rectime, censrec) ~ group, data = flexsurv::bc, dist = "gengamma"))
}
